Add files via upload
On index.php page user can see all forums with stats for each forum (how many topics and posts has it, date of the last post and who created it). Each forum name is linked and by clicking on it, user will be redirected to forum.php page where all topics inside that specific forum will be listed with related stats (number of posts and date of the last post). Pagination of both topic.php and forum.php page is set up to show 3 of it on the same page. 
By clicking on Log in, user will be redirected to login-form.php page where there are two inputs- one for email address and the other for password. This form will be checked on user.php form and redirect to index.php if the entered credentials are ok. if not it will report an error. 
By clicking on Sign up, user will be asked to enter data for signing up and if the email address isn't used already and no one has the same user name, new user will be entered in the forum database. 
When user is signed in, he has an option of staring a new forum or topic and index.php and forum.php pages. Also, he can change/delete only his own post.
If user types anything in Search input, it will be redirected to search.php where results related to the searched word(s) will .be displayed, if any.